+++ Libft/libft_tester/tests/ft_strlcpy_test.c
@@ -6,11 +6,12 @@
 /*   By: npentini <npentini@student.42abudhabi.ae>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/01 22:27:57 by npentini          #+#    #+#             */
-/*   Updated: 2023/07/07 18:32:08 by npentini         ###   ########.fr       */
+/*   Updated: 2023/07/07 22:23:04 by npentini         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "testutils.h"
+#include <bsd/string.h>
 
 void	ft_strlcpy_test(int cols, int rows, int argc, char **argv)
 {
@@ -20,7 +21,7 @@ void	ft_strlcpy_test(int cols, int rows, int argc, char **argv)
 	char					*dest1;
 	char					*dest2;
 	int						x;
-	int						p;
+	size_t					p;
 	size_t					size;
 
 	info = get_function_info(argv[0]);
